Use same swig input file for golang as race-core
This ensures that the unique ID added onto generated symbol names
matches those of the swig golang bindings in race-core, allowing plugins
built against race-core to be used for raceboat. Since headers have
moved, this means that raceboat has to set the particular include
subdirectories so that the headers can be found by swig.